Fixed bugs and changed roles field
Fixed a bug with the name field: when sending a name in a POST request, a new name instance was created - which caused issues when the name already exists.
The updated code searches for the name in the DB and creates a new one only if it wasn't able to find one.

Modified the role field: Changed the `roles` field in the `Customer` model to be ManyToMany field. So now we can create new customers it references roles.